+++ HISTORY.rst
@@ -24,6 +24,18 @@ latest (unreleased)
 **Features and Improvements**
 
 **Bugfixes**
+
+**Build**
+
+**Documentation**
+
+
+9.41.2 (2018-02-07)
++++++++++++++++++++
+
+**Features and Improvements**
+
+**Bugfixes**
 * BIZ-747 : Missing recompute taxes* BIZ-747
 **Build**